svg/text/parsing/text-decoration-stroke-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing text-decoration-stroke with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#TextDecorationStrokeProperty" rel="help"/><h:meta content="text-decoration-stroke supports only the paint grammar 'none | color | url [none | color]? | context-fill | context-stroke'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("text-decoration-stroke", "auto");
test_invalid_value("text-decoration-stroke", "none red");
test_invalid_value("text-decoration-stroke", 'none url("https://example.com/")');
test_invalid_value("text-decoration-stroke", 'red url("https://example.com/")');
test_invalid_value("text-decoration-stroke", 'url("https://example.com/") none red');

  ]]></script></svg>
